open_api_test.py
import os

from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()


API_KEY = os.getenv('OPENAI_API_KEY')

client = OpenAI(
  api_key=API_KEY
)

completion = client.chat.completions.create(
  model="gpt-4o-mini",
  store=True,
  messages=[
    {
      "role": "user", 
      "content": "write a haiku about ai",
    }
  ]
)

print(completion.choices[0].message);
