requirements.txt
crewai==0.141.0
httpx==0.28.1
httpx-sse==0.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
pandas==2.3.1
python-dotenv==1.1.1